+++ lib/src/datasources/datasource_ncc_delete.dart
@@ -33,12 +33,5 @@ abstract base class DatasourceNccDelete<
     required Map<String, String>? requestHeaders,
     required Uri? requestUri,
     Object? requestBody,
-  }) {
-    return super.checkInformation(
-      requestResponse: requestResponse,
-      requestHeaders: requestHeaders,
-      requestUri: requestUri,
-      requestBody: requestBody,
-    );
-  }
+  });
 }