+++ lib/src/datasources/datasource_ncc_get.dart
@@ -32,12 +32,5 @@ abstract base class DatasourceNccGet<
     required Map<String, String>? requestHeaders,
     required Uri? requestUri,
     Object? requestBody,
-  }) {
-    return super.checkInformation(
-      requestResponse: requestResponse,
-      requestHeaders: requestHeaders,
-      requestUri: requestUri,
-      requestBody: requestBody,
-    );
-  }
+  });
 }