+++ test/mocks/datasource/http_datasources/test_patch_datasource.dart
@@ -1,6 +1,6 @@
 part of '../implementation_datasource.dart';
 
-final class TestHttpPatchDataSource extends DatasourcePatchHttp<MockModel>
+final class TestHttpPatchDataSource extends DatasourceHttpPatch<MockModel>
     with ImplementationDatasource {
   TestHttpPatchDataSource({required super.driver});