+++ test/mocks/datasource/session_datasources/test_get_datasource.dart
@@ -1,6 +1,6 @@
 part of '../implementation_datasource.dart';
 
-final class TestSessionGetDataSource extends DatasourceGetSession<MockModel>
+final class TestSessionGetDataSource extends DatasourceSessionGet<MockModel>
     with ImplementationDatasource {
   TestSessionGetDataSource({required super.driver});