+++ test/mocks/datasource/session_datasources/test_put_datasource.dart
@@ -1,6 +1,6 @@
 part of '../implementation_datasource.dart';
 
-final class TestSessionPutDataSource extends DatasourcePutSession<MockModel>
+final class TestSessionPutDataSource extends DatasourceSessionPut<MockModel>
     with ImplementationDatasource {
   TestSessionPutDataSource({required super.driver});